(minor) rename the interface 'searchResults' to 'search' and add drop shadow to the dropdown in the compendia
