+++ src/api/index.js
@@ -3,7 +3,7 @@ import { downloader, processor } from './interfaces/jobs'
 import experiments from './interfaces/experiments'
 import originalFiles from './interfaces/originalFiles'
 import samples from './interfaces/samples'
-import searchResults from './interfaces/searchResults'
+import search from './interfaces/search'
 
 export const api = {
   computedFiles,
@@ -12,5 +12,5 @@ export const api = {
   originalFiles,
   processor,
   samples,
-  searchResults
+  search
 }
